add pages and session states
