+++ Dashboard/Pages/District_Map.py
@@ -1,0 +1,12 @@
+from streamlit_folium import folium_static
+import streamlit as st
+import pandas as pd
+import visualizations as v
+
+def show_folium_map():
+    st.title('Singapore District Map')
+    m = v.plot_price_per_district()
+    # Display the map
+    folium_static(m)
+    
+show_folium_map()
